battlescape unit progres and fixes
- display unit selection arrows
- proper strategy draw order in battle,
- fixed some travis warnings,
- provide methods to change button sound

+++ game/state/battle/battleforces.cpp
@@ -10,14 +10,14 @@ BattleForces::BattleForces() : squads(6) {};
 
 bool BattleForces::insert(unsigned squad, sp<BattleUnit> unit)
 {
-	if (squads[squad].getNumUnits() == 6 && unit->squadNumber != squad)
+	if (squads[squad].getNumUnits() == 6 && unit->squadNumber != (int)squad)
 		return false;
 	return insertAt(squad, squads[squad].getNumUnits(), unit);
 }
 
 bool BattleForces::insertAt(unsigned squad, unsigned position, sp<BattleUnit> unit)
 {
-	if (squads[squad].getNumUnits() == 6 && unit->squadNumber != squad)
+	if (squads[squad].getNumUnits() == 6 && unit->squadNumber != (int)squad)
 		return false;
 	if (unit->squadNumber != -1)
 	{
